python/telegram-bot/requirements.txt
anyio==3.6.2
asyncio==3.4.3
black==23.3.0
certifi==2022.12.7
click==8.1.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.4.0
python-dotenv==1.0.0
python-telegram-bot==20.2
rfc3986==1.5.0
sniffio==1.3.0
